HIBERNATE FIX

+++ src/main/java/pl/umk/mat/planner/lecturer/LecturerRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface LecturerRepository extends JpaRepository<Lecturer, Integer> {
+public interface LecturerRepository extends JpaRepository<Lecturer, Long> {
 }
